Improve DASH segment caching
Caching key type changed from segment (which is a different object for every representation) to integer (the segment sequence is similar in each representation [at least for the current test videos; could be different though]), so no matter which representation's segment is cached, it can be taken from the cache by with the same key. Because of this change, the representation can now change at precaching time instead of directly at playback time (through a blocking download), resulting in much smoother playback.