Improve video resolution change handling
This changes the surface size from being fixed to the resolution of the video to taking the size that the UI permits. Because of this change, the resizing of the fixed surface is no longer visible, resulting in smoother DASH representation changes. It also changes the processing resolution / texture size of the GLES view, which is now also the taken screen size instead of the video size. This speeds up shader processing if the native video resolution is higher that the screen space, but slows it down if the video's resolution is smaller; optimally, it would be the screen space size or video size, whatever is smaller.