Improve DASH representation switching and fix seeking
The reinitialization of the codec for the new format is now getting deferred until all frames of the old representation have been rendered. Before, these frames were lost and skipped, resulting in a jittery representation switch. This also fixes seeking into representations that differ from the active representation. There is still a bit of a lag when switching representations, resulting from the MediaCodec initialization which takes some time. The video resolution changed message is also getting deferred until the first frame of the new representation is getting rendered.