Fix DASH seeking issues with multiple shortly following requests
This problem appeared in the demo app because the Android MediaController issues seek requests whenever a seek happens on the timeline slider. When the user touches the slider to seek and moves the finger a tiny bit while touching, multiple (mostly two) seeks are requested... these shortly following consecutive seeks triggered the problem that this commit solves, which led to the seek operation not being carried out as the user expected it.