Add stopPlayback() methods to the video views
This also changes the stop() behavior in MediaPlayer, which is now blocking until the playback thread is stopped. This can make a difference when stopping while a seek is in progress, because it now waits until the seek call returns, which can take some time with online streams on a slow connection.