Fix OnSeekListener being called from wrong thread
The OnSeekListener added in commit ed303b561 was called from the seekTo method, which was also called internally by the PlaybackThread of the MediaPlayer on some occasions, leading to the listener being called by an unexpected thread. This commit splits the onSeek method into a private method for internal MediaPlayer calls that does not call any listeners, and a public method that calls the listener.