Fix seeking from the end of the input stream
When seeking to the end of a stream and the end of the input stream is reached, the mVideoInputEos flag gets set to true. A following seek then resulted in an infinite loop of the media player waiting for a decoded frame, because due to the flag, no more sample data was read from the extractor and fed to the decoder, and the decoder therefore never emitted a decoded frame. Fixes GitHub issue #9.