Fix that fucking handler delay problem
Been searching for months for the source of the problem that delayed messages in the mediaplayer handler took 5x longer than specified, leading to bad playback performance and stuttering. Turns out it's the AsyncTask's low priority that affects all threads created from the async task, no matter which priority they are given.