Take load from decoder by pausing the loop while no output is available
Before, the decoder was looping and requesting a decoded video frame until it became available, wasting processing power while the decoder needed time to decode the frame. Now we just ask for the frame once and then delay the next query for some time, putting the decoder loop to sleep.