Update audio output format when codec output format changes
On some devices, the MediaCodec decoder outputs stereo audio data for mono input data, which results in slowmotion audio playback when the AudioTrack is configured with the MediaFormat from the extractor (mono) instead of the MediaFormat of the decoder. By reinitializing the AudioTrack when the codec output MediaFormat changes, audio playback should always play at the correct speed.