Improve releasing of extractors
Extractors weren't released if a player was released which had a data source set but was not prepared. Only if the player was prepared, the thread that released the extractors was created and destroyed. Now the extractors are released in any case. Not releasing the extractors was not necessarily an issue, but especially for DASH it could happen that downloads continued to run, and file handles were left open.