Operate on a local player reference inside the thread
Operating on the global reference led to invalid call sequences (e.g. release, prepareAsync) because the reference could change (e.g. due to releasing or opening new video) while the thread was doing things. The thread now always works on the same reference, detects if the reference is outdated and released the (old) player in such a case.

This probably fixes GitHub issue #70.